feat: finalizando a criação dos models

+++ src/models/Filme.js
@@ -22,7 +22,7 @@ const Filme = sequelize.define(
       allowNull: false,
     },
     duracao: {
-      type: DataTypes.NUMERIC,
+      type: DataTypes.INTEGER,
       allowNull: false,
     },
   },
